refactor: migrate cli logic to ts from js
- write whole command building and actions logic of commander in
src/index.ts for future scalability and type safety
- bin/cli.js now only contains importing this src/index.ts compiled
lib/index.js file.
